macOS: disallow run from dmg volume (#238)
* Stop macOS users from running the app from the DMG

If a user tries to run the app from the mounted DMG
installer, it will exhibit similar issues to a quarantined application:
namely, read-only characteristics, which can prevent sign-in and other
functionality from working properly.

This patch alters the quarantine check to _also_ check for if they're
running from the mounted DMG volume. If it is, we'll display a similar
message to the user indicating what they should do.