Revert behavior of wxStrToStr.
Looking at this, I think it was changed when getting Externals/wxWidgets
updated to 3.1.4, but this particular method didn't need to stay
changed. For users who run in different locales, I suspect it's causing
issues with folder paths on Windows that contain special characters.

Reverting the behavior does not break the build and my initial testing
shows no issues. I believe this went undetected during testing due to
the aforementioned locale issue, with none of us being in a different
locale.