Print assert message as string, not as format (#294)
The 3rd argument to _assert_msg_ is passed to MsgAlert, and will be
interpreted as a format string. This is not intended behavior for the
msg passed here, so we fix the print by formatting msg as a regular C
string. This also fixes a compiler error (-Werror=format-security) when
building with GCC.