macOS: Default to Vulkan for rendering (#306)
* Change macOS to default to Vulkan in 10.14+.

This mirrors a change in mainline Dolphin, so should be relatively safe
to do. Slippi's MoltenVK setup is a bit older than modern Dolphin's, but
most users still see a jump in performance in comparison to Apple's
broken OpenGL API.

High Sierra remains defaulted to OpenGL, as Metal on that platform lacks
critical pieces that Dolphin requires.