Adds a macOS-specific `AllocateExecutableMemory()`
I was thinking about the 2GB Executable Memory error that users
experience, and decided to look into it again. I realized that this
function actually predates macOS Catalina, and since it's building
against an older SDK it'd have been going the "best effort" route due to
no support for `MAP_32BIT`.

However, Catalina actually introduced `MAP_32BIT` and made it more
widely available in `10.15.4`. This is one of the differences in how
this block of code executes (e.g, Linux vs macOS). `MAP_32BIT` can be
used to request allocation below the 4GB boundary, which is what we're
after here.

The block itself is just separated entirely so as to not muck with the
rest of the platforms, and I condensed the `MAP_JIT` check into here as
well for brevity.

I've tested the built version of this on 10.15.7 and 11.5 and not hit
issues, but this should probably still be tested further.