fix: prevent early game inputs from being ignored
This was highlighted by the tiebreaks in ranked where two games start very quickly back to back but it was possible for old remote inputs to linger on a new game and cause new inputs from the start of the game to get acked immediately but not actually used.

This will hopefully make it so that they don't get acked, the remote queue will get cleaned up, and then the inputs from the new game will be correctly received.