fix: prevent potential crashes
one crash here was possible when a new game started but some remote inputs lingered, causing an insert to have a source iterator that started after the iterator end.

The other crash happened because I switched to ptrs in orderedSelections such that lps and rps values are properly updated. After doing this though, the vector had to be the same length as the number of players in the game instead of always being 4.