chore: add modified injection list
for now since this one removed a bunch of codes, I want to keep the old one in the folder as well, otherwise replays from 2.5.0 and 2.5.1 will load a bunch of dynamic gecko codes that shouldn't be loaded.

We probably need to come up with a better way to manage this, perhaps a merging script?