Shim SlippiDirectCodes to Rust port.
This changes out the internals of the `SlippiDirectCodes` class, wherein
it'll now just silently call over to the Rust side. This keeps the
changes minimal for now as we continue to migrate things out.

There is a small change here to `FileUtil` to have it return a folder
path rather than a file path; the Rust layer will extend that path with
whatever it needs (e.g `user.json`).

This compiles but has not been fully tested/vetted yet. Building this
branch currently requires the corresponding branch from
slippi-rust-extensions.